requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
requests>=2.26.0
python-dotenv>=0.19.0
pydantic-settings==2.0.1
pydantic[email]
openpyxl